+++ src/sagemaker/mxnet/model.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"). You
 # may not use this file except in compliance with the License. A copy of
@@ -13,6 +13,7 @@
 import sagemaker
 from sagemaker.fw_utils import create_image_uri
 from sagemaker.model import FrameworkModel, MODEL_SERVER_WORKERS_PARAM_NAME
+from sagemaker.mxnet.defaults import MXNET_VERSION
 from sagemaker.predictor import RealTimePredictor, json_serializer, json_deserializer
 from sagemaker.utils import name_from_image
 
@@ -20,7 +21,8 @@ from sagemaker.utils import name_from_image
 class MXNetPredictor(RealTimePredictor):
     """A RealTimePredictor for inference against MXNet Endpoints.
 
-    This is able to serialize Python lists and numpy arrays to multidimensional tensors for MXNet inference."""
+    This is able to serialize Python lists, dictionaries, and numpy arrays to multidimensional tensors for MXNet
+    inference."""
 
     def __init__(self, endpoint_name, sagemaker_session=None):
         """Initialize an ``MXNetPredictor``.
@@ -39,7 +41,7 @@ class MXNetModel(FrameworkModel):
 
     __framework_name__ = 'mxnet'
 
-    def __init__(self, model_data, role, entry_point, image=None, py_version='py2',
+    def __init__(self, model_data, role, entry_point, image=None, py_version='py2', framework_version=MXNET_VERSION,
                  predictor_cls=MXNetPredictor, model_server_workers=None, **kwargs):
         """Initialize an MXNetModel.
 
@@ -53,6 +55,7 @@ class MXNetModel(FrameworkModel):
                 as the entry point to model hosting. This should be compatible with either Python 2.7 or Python 3.5.
             image (str): A Docker image URI (default: None). If not specified, a default image for MXNet will be used.
             py_version (str): Python version you want to use for executing your model training code (default: 'py2').
+            framework_version (str): MXNet version you want to use for executing your model training code.
             predictor_cls (callable[str, sagemaker.session.Session]): A function to call to create a predictor
                 with an endpoint name and SageMaker ``Session``. If specified, ``deploy()`` returns the result of
                 invoking this function on the created endpoint name.
@@ -63,6 +66,7 @@ class MXNetModel(FrameworkModel):
         super(MXNetModel, self).__init__(model_data, image, role, entry_point, predictor_cls=predictor_cls,
                                          **kwargs)
         self.py_version = py_version
+        self.framework_version = framework_version
         self.model_server_workers = model_server_workers
 
     def prepare_container_def(self, instance_type):
@@ -77,8 +81,8 @@ class MXNetModel(FrameworkModel):
         deploy_image = self.image
         if not deploy_image:
             region_name = self.sagemaker_session.boto_session.region_name
-            deploy_image = create_image_uri(region_name, self.__framework_name__, instance_type, self.py_version,
-                                            sagemaker.mxnet.DOCKER_TAG)
+            deploy_image = create_image_uri(region_name, self.__framework_name__, instance_type,
+                                            self.framework_version, self.py_version)
         deploy_key_prefix = self.key_prefix or self.name or name_from_image(deploy_image)
         self._upload_code(deploy_key_prefix)
         deploy_env = dict(self.env)