+++ tests/integ/test_mxnet_train.py
@@ -19,6 +19,7 @@ import pytest
 from sagemaker import Session
 from sagemaker.mxnet.estimator import MXNet
 from sagemaker.mxnet.model import MXNetModel
+from sagemaker.utils import sagemaker_timestamp
 
 from tests.integ import DATA_DIR, REGION
 from tests.integ.timeout import timeout, timeout_and_delete_endpoint_by_name
@@ -30,12 +31,12 @@ def sagemaker_session():
 
 
 @pytest.fixture(scope='module')
-def mxnet_training_job(sagemaker_session):
+def mxnet_training_job(sagemaker_session, mxnet_full_version):
     with timeout(minutes=15):
         script_path = os.path.join(DATA_DIR, 'mxnet_mnist', 'mnist.py')
         data_path = os.path.join(DATA_DIR, 'mxnet_mnist')
 
-        mx = MXNet(entry_point=script_path, role='SageMakerRole',
+        mx = MXNet(entry_point=script_path, role='SageMakerRole', framework_version=mxnet_full_version,
                    train_instance_count=1, train_instance_type='ml.c4.xlarge',
                    sagemaker_session=sagemaker_session)
 
@@ -49,7 +50,7 @@ def mxnet_training_job(sagemaker_session):
 
 
 def test_attach_deploy(mxnet_training_job, sagemaker_session):
-    endpoint_name = 'test-mxnet-attach-deploy-{}'.format(int(time.time()))
+    endpoint_name = 'test-mxnet-attach-deploy-{}'.format(sagemaker_timestamp())
 
     with timeout_and_delete_endpoint_by_name(endpoint_name, sagemaker_session, minutes=20):
         estimator = MXNet.attach(mxnet_training_job, sagemaker_session=sagemaker_session)
@@ -58,8 +59,40 @@ def test_attach_deploy(mxnet_training_job, sagemaker_session):
         predictor.predict(data)
 
 
+def test_async_fit(sagemaker_session, mxnet_full_version):
+
+    training_job_name = ""
+    endpoint_name = 'test-mxnet-attach-deploy-{}'.format(sagemaker_timestamp())
+
+    with timeout(minutes=5):
+        script_path = os.path.join(DATA_DIR, 'mxnet_mnist', 'mnist.py')
+        data_path = os.path.join(DATA_DIR, 'mxnet_mnist')
+
+        mx = MXNet(entry_point=script_path, role='SageMakerRole', framework_version=mxnet_full_version,
+                   train_instance_count=1, train_instance_type='ml.c4.xlarge',
+                   sagemaker_session=sagemaker_session)
+
+        train_input = mx.sagemaker_session.upload_data(path=os.path.join(data_path, 'train'),
+                                                       key_prefix='integ-test-data/mxnet_mnist/train')
+        test_input = mx.sagemaker_session.upload_data(path=os.path.join(data_path, 'test'),
+                                                      key_prefix='integ-test-data/mxnet_mnist/test')
+
+        mx.fit({'train': train_input, 'test': test_input}, wait=False)
+        training_job_name = mx.latest_training_job.name
+
+        print("Waiting to re-attach to the training job: %s" % training_job_name)
+        time.sleep(20)
+
+    with timeout_and_delete_endpoint_by_name(endpoint_name, sagemaker_session, minutes=35):
+        print("Re-attaching now to: %s" % training_job_name)
+        estimator = MXNet.attach(training_job_name=training_job_name, sagemaker_session=sagemaker_session)
+        predictor = estimator.deploy(1, 'ml.m4.xlarge', endpoint_name=endpoint_name)
+        data = numpy.zeros(shape=(1, 1, 28, 28))
+        predictor.predict(data)
+
+
 def test_deploy_model(mxnet_training_job, sagemaker_session):
-    endpoint_name = 'test-mxnet-deploy-model-{}'.format(int(time.time()))
+    endpoint_name = 'test-mxnet-deploy-model-{}'.format(sagemaker_timestamp())
 
     with timeout_and_delete_endpoint_by_name(endpoint_name, sagemaker_session, minutes=20):
         desc = sagemaker_session.sagemaker_client.describe_training_job(TrainingJobName=mxnet_training_job)
@@ -72,12 +105,12 @@ def test_deploy_model(mxnet_training_job, sagemaker_session):
         predictor.predict(data)
 
 
-def test_failed_training_job(sagemaker_session):
+def test_failed_training_job(sagemaker_session, mxnet_full_version):
     with timeout(minutes=15):
         script_path = os.path.join(DATA_DIR, 'mxnet_mnist', 'failure_script.py')
         data_path = os.path.join(DATA_DIR, 'mxnet_mnist')
 
-        mx = MXNet(entry_point=script_path, role='SageMakerRole',
+        mx = MXNet(entry_point=script_path, role='SageMakerRole', framework_version=mxnet_full_version,
                    train_instance_count=1, train_instance_type='ml.c4.xlarge',
                    sagemaker_session=sagemaker_session)