+++ CHANGELOG.rst
@@ -2,11 +2,52 @@
 CHANGELOG
 =========
 
+1.1.2
+=======
+
+* bug-fix: AmazonEstimators: do not call create bucket if data location is provided
+
+1.1.1
+========
+
+* feature: Estimators: add ``requirements.txt`` support for TensorFlow
+
+
+1.1.0
+=====
+
+* feature: Estimators: add support for TensorFlow-1.5.0
+* feature: Estimators: add support for MXNet-1.0.0
+* feature: Tests: use ``sagemaker_timestamp`` when creating endpoint names in integration tests
+* feature: Session: print out billable seconds after training completes
+* bug-fix: Estimators: fix LinearLearner and add unit tests
+* bug-fix: Tests: fix timeouts for PCA async integration test
+* feature: Predictors: allow ``predictor.predict()`` in the JSON serializer to accept dictionaries
+
+1.0.4
+=====
+
+* feature: Estimators: add support for Amazon Neural Topic Model(NTM) algorithm
+* feature: Documentation: fix description of an argument of sagemaker.session.train
+* feature: Documentation: add FM and LDA to the documentation
+* feature: Estimators: add support for async fit
+* bug-fix: Estimators: fix estimator role expansion
+
+1.0.3
+=====
+
+* feature: Estimators: add support for Amazon LDA algorithm
+* feature: Hyperparameters: add data_type to hyperparameters
+* feature: Documentation: update TensorFlow examples following API change
+* feature: Session: support multi-part uploads
+* feature: add new SageMaker CLI
+
+
 1.0.2
 =====
 
 * feature: Estimators: add support for Amazon FactorizationMachines algorithm
-* feature: Session: Correctly handle TooManyBuckets error_code in default_bucket method
+* feature: Session: correctly handle TooManyBuckets error_code in default_bucket method
 * feature: Tests: add training failure tests for TF and MXNet
 * feature: Documentation: show how to make predictions against existing endpoint
 * feature: Estimators: implement write_spmatrix_to_sparse_tensor to support any scipy.sparse matrix
@@ -27,4 +68,3 @@ CHANGELOG
 =====
 
 * Initial commit
-